requirements.txt
asgiref==3.9.1
diff-match-patch==20241021
Django==5.2.5
django-cors-headers==4.7.0
django-filter==25.1
django-import-export==4.3.9
djangorestframework==3.16.1
Faker==37.6.0
pillow==11.3.0
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
